fix the layout of absolute and relative position
