+++ src/Components/Intro.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import BackgroundCircle from "./BackgroundCircle";
+import BackgroundCircles from "./BackgroundCircles";
 import imageCartoonised from "../Resouce/personal-catoonized-no-bg.png";
 import Image from "next/image";
 
@@ -7,13 +8,8 @@ type Props = {};
 
 export default function Intro({}: Props) {
   return (
-    <div className="relative flex flex-col items-center justify-center text-center overflow-hidden ">
-      <BackgroundCircle />
-      {/* <Image
-              src={imageCartoonised}
-              alt="my photo"
-              className="relative w-40 h-40 bg-pink-200 rounded-full object-cover items-center"
-            /> */}
+    <div className="flex flex-col items-center justify-center text-center overflow-hidden ">
+      <BackgroundCircles />
     </div>
   );
 }
